plataforma_automacao_ia_generativa/app_saas/backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary # Para PostgreSQL
pydantic
pydantic-settings
python-dotenv
google-generativeai